two conns

+++ database.py
@@ -14,7 +14,7 @@ from psycopg2.pool import SimpleConnectionPool as simPool
 up.uses_netloc.append("postgres")
 url = up.urlparse(os.environ["DATABASE_URL"])
 
-pool = simPool(1, 5, database=url.path[1:], user=url.username,
+pool = simPool(0, 2, database=url.path[1:], user=url.username,
             password=url.password, host=url.hostname, port=url.port)
 
 # gets clubs corresponding matching the given phrase and/or tags
@@ -22,6 +22,11 @@ def get_clubs(clubquery, tags):
 
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            pool.closeall()
+            return get_clubs(clubquery, tags)
+
+        print("size:", pool.__sizeof__)
 
         with conn.cursor() as cur:
 
@@ -108,7 +113,8 @@ def get_clubs(clubquery, tags):
         print("__________database.py:", ex)
         return "server get_clubs"
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 # get club details of a selected club
@@ -116,6 +122,11 @@ def database_get_info(clubid):
 
     try:
         conn = pool.getconn()
+
+        print("size:", pool.__sizeof__)
+        if 'conn' not in locals():
+            pool.closeall()
+            return database_get_info(clubid)
 
         with conn.cursor() as cur:
             script = "select groupname, mission, goals, groupemail, instagram, youtube, imlink from clubs WHERE id=%s"
@@ -129,11 +140,12 @@ def database_get_info(clubid):
             return info
 
     except psycopg2.OperationalError as ex:
-        print("__________database.py get clubs: ", ex)
+        print("__________database.py get info: ", ex)
 
         return "server, get_info"
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
     ############ catch general exception?
 
 
@@ -141,6 +153,10 @@ def database_get_info(clubid):
 def get_tags():
     try:
         conn = pool.getconn()
+
+        print("size:", pool.__sizeof__())
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
 
         with conn.cursor() as cur:
 
@@ -160,13 +176,16 @@ def get_tags():
         return "server get_tags"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 # get tags linked to a club
 def get_club_tags(clubid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "select tag from tags WHERE id=%s order by tag asc"
@@ -185,13 +204,16 @@ def get_club_tags(clubid):
         return "server get_tags"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 # subscribes user to club
 def add_sub(user, clubid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             # check for existence of row (may not be necessary)
@@ -208,13 +230,17 @@ def add_sub(user, clubid):
         return "server, add_sub"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            conn.commit()
+            pool.putconn(conn)
 
 
 # unsubscribes user from club
 def remove_sub(user, clubid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "select * from subscriptions where netid=%s and id=%s"
@@ -234,13 +260,17 @@ def remove_sub(user, clubid):
         return "server, remove_sub"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            conn.commit()
+            pool.putconn(conn)
 
 
 # subscribes user to tag
 def add_sub_tag(user, tag):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
 
         with conn.cursor() as cur:
 
@@ -260,6 +290,9 @@ def add_sub_tag(user, tag):
     except Exception as ex:
         print("__________database.py add sub tag:", ex)
         return "server, add_sub_tag"
+    finally:
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 # unsubscribes user from tag
@@ -267,6 +300,8 @@ def remove_sub_tag(user, tag):
     retval = None
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "select distinct tags.id from tags, subscriptions where tag=%s and netid=%s and tags.id=subscriptions.id"
@@ -281,20 +316,26 @@ def remove_sub_tag(user, tag):
                     remove_sub(user, clubid)
                 clubid = cur.fetchone()
 
+        return retval
+
     except Exception as ex:
         print("__________database.py remove sub tag:", ex)
         return "server, remove_sub_tag"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
-    return retval
+    
 
 
 # check if a user is subscribed to a club
 def is_subbed(netid, clubid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            pool.closeall()
+            return is_subbed(netid, clubid)
         with conn.cursor() as cur:
 
             script = "select netid from subscriptions WHERE netid=%s AND id=%s"
@@ -313,12 +354,16 @@ def is_subbed(netid, clubid):
         return "server, is_subbed"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 def get_user_sub_tags(netid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            pool.closeall()
+            return get_user_sub_tags(netid)
         with conn.cursor() as cur:
 
             script = "select distinct tag from tags WHERE id in (select id from subscriptions where netid=%s) order by tag asc"
@@ -339,7 +384,8 @@ def get_user_sub_tags(netid):
         return "server, get_user_sub_tags"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 # gets all announcements from club with given clubid and returns
@@ -350,6 +396,9 @@ def get_club_announcements(clubid):
 
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            pool.closeall()
+            return get_club_announcements(clubid)
         with conn.cursor() as cur:
 
             cur.execute(script, [clubid])
@@ -359,10 +408,11 @@ def get_club_announcements(clubid):
 
     except Exception as ex:
         print("__________database.py get club announcements: ", ex)
-        return "server, get_club_announcements"
+        return ["server, get_club_announcements","",""]
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 ###### -------------- Profile Information -------------------- ######
@@ -371,6 +421,8 @@ def get_club_announcements(clubid):
 def get_subs(netid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "select groupname, clubs.id from subscriptions, clubs WHERE netid=%s AND clubs.id=subscriptions.id"
@@ -389,13 +441,16 @@ def get_subs(netid):
         return "server, get_subs"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 # return clubids and clubnames for clubs where netid is an officer
 def get_officerships(netid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "select groupname, clubs.id from officers, clubs"
@@ -416,12 +471,15 @@ def get_officerships(netid):
         return "server, get_officerships"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 # return clubids and clubnames for clubs where netid is an officer
 def get_club_officers(clubid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "select netid from officers WHERE clubid=%s"
@@ -446,7 +504,8 @@ def get_club_officers(clubid):
         return "server, get_club_officers"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 
@@ -456,6 +515,8 @@ def get_club_officers(clubid):
 def verify_officer(netid, clubid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "select * from officers where netid=%s and clubid=%s"
@@ -471,12 +532,15 @@ def verify_officer(netid, clubid):
         return "server, verify_officer"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 def add_officer(netid, clubid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "insert into officers values (%s, %s)"
@@ -493,13 +557,17 @@ def add_officer(netid, clubid):
         return "server, add_officer"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            conn.commit()
+            pool.putconn(conn)
 
 
 def remove_officer(netid, clubid):
     print("in remove_officer")
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "delete from officers where netid=%s and clubid=%s"
@@ -514,7 +582,9 @@ def remove_officer(netid, clubid):
         return "server, remove_officer"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            conn.commit()
+            pool.putconn(conn)
 
 
 # send an announcement to a given club
@@ -523,6 +593,8 @@ def send_announcement(clubid, announcement, officer):
 
         # update the database of announcements
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "insert into announcements VALUES (%s, %s, CURRENT_TIMESTAMP, %s)"
@@ -538,13 +610,17 @@ def send_announcement(clubid, announcement, officer):
         return "server, send_announcements"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            conn.commit()
+            pool.putconn(conn)
 
 
 # get all subscribers for a club
 def get_subscribers(clubid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "select netid from subscriptions where id=%s"
@@ -564,7 +640,8 @@ def get_subscribers(clubid):
         return "server, get_subscribers"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 
 # updates club info in database
@@ -581,6 +658,8 @@ def update_club_info(clubid, instagram=None, youtube=None, email=None,
 
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
             cur.execute(script, args+[clubid])
             conn.commit()
@@ -590,13 +669,17 @@ def update_club_info(clubid, instagram=None, youtube=None, email=None,
         return "server, update_club_info"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            conn.commit()
+            pool.putconn(conn)
 
 
 # get clubname from clubid
 def get_clubname(clubid):
     try:
         conn = pool.getconn()
+        if 'conn' not in locals():
+            raise psycopg2.OperationalError()
         with conn.cursor() as cur:
 
             script = "select MAX(id) from clubs"
@@ -616,7 +699,8 @@ def get_clubname(clubid):
         return "server, get_clubname"
 
     finally:
-        pool.putconn(conn)
+        if 'conn' in locals():
+            pool.putconn(conn)
 
 ################
 # getting user infos from netid
